allow simulation mode; invoke basicft8.py

+++ utils/ft8.py
@@ -1,3 +1,4 @@
+import sys
 import socket
 import time
 import subprocess
@@ -7,8 +8,18 @@ import string
 cq_syms=[2, 5, 6, 0, 4, 1, 3, 5, 7, 3, 0, 0, 0, 3, 3, 6, 2, 7, 4, 3, 4, 0, 1, 1, 2, 2, 2, 1, 5, 6, 3, 2, 3, 1, 2, 0, 2, 5, 6, 0, 4, 1, 3, 7, 6, 4, 0, 4, 0, 6, 1, 4, 7, 6, 6, 2, 7, 2, 6, 0, 2, 0, 3, 7, 7, 2, 5, 0, 6, 0, 1, 6, 2, 5, 6, 0, 4, 1, 3]
 test_syms=[2, 5, 6, 0, 4, 1, 3, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 2, 5, 6, 0, 4, 1, 3, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 9, 8, 2, 5, 6, 0, 4, 1, 3]
 
+#{0: 1500.0,
+# 1: 1506.25,
+# 2: 1512.5,
+# 3: 1518.75,
+# 4: 1525.0,
+# 5: 1531.25,
+# 6: 1537.5,
+# 7: 1543.75}
+
 g_tones={}
 old_freq_offset=0.0
+old_twice_dac_offset=0.0
 
 def init_tones(basefreq):
     for i in range(8):
@@ -16,10 +27,12 @@ def init_tones(basefreq):
 
 def freq_to_dac(freq):
     global old_freq_offset
+    global old_twice_dac_offset
+
     base_f=1500
-    base_dac=2611
-    base_dac=2601
-    base_dac=2598
+    #base_dac=2611
+    #base_dac=2601
+    base_dac=2605
     twice_base_dac=base_dac*2
 
     hz_per_count = 1.046
@@ -27,9 +40,30 @@ def freq_to_dac(freq):
     freq_offset=freq-base_f
 
     diff_offset = freq_offset - old_freq_offset
-    
-    twice_dac_offset=(2*(freq_offset+(diff_offset/4.2)))/hz_per_count
 
+    sqf = diff_offset/16
+    print sqf
+    if diff_offset>=0:
+        square_offset = sqf*sqf
+    else:
+        square_offset = -1*sqf*sqf
+
+    # 4.4 and 16 look ok
+    twice_dac_offset=(2*(freq_offset+(diff_offset/4.4)+square_offset))/hz_per_count
+    #twice_dac_offset=(2*freq_offset)/hz_per_count
+
+    #print twice_dac_offset - old_twice_dac_offset
+
+    if False:
+        boundary=10
+        fudge=5
+        if diff_offset > 0:
+            if (twice_dac_offset - old_twice_dac_offset) < boundary:
+                twice_dac_offset += fudge
+        else:
+            if (old_twice_dac_offset - twice_dac_offset) < boundary:
+                twice_dac_offset -= fudge
+            
     twice_dac_offset_int = int(round(twice_dac_offset))
 
     twice_dac_val = twice_dac_offset_int + twice_base_dac
@@ -45,6 +79,7 @@ def freq_to_dac(freq):
     #print result
 
     old_freq_offset = freq_offset
+    old_twice_dac_offset = twice_dac_offset
     
     return result
     
@@ -68,53 +103,65 @@ if __name__ == "__main__":
     init_tones(1500)
     print g_tones
 
-    send_msg("ft8-txon")
-    time.sleep(1)
-    zero = "D2A26"
-    send_msg(zero)
-
     print len(test_syms)
     test_syms2 = [ x if x!=8 else 7 for x in test_syms ]
-    test_syms3 = [ x if x!=9 else 2 for x in test_syms2 ]
+    test_syms3 = [ x if x!=9 else 0 for x in test_syms2 ]
     print test_syms3
+
+    zero = "D2A2D"
+
+    sim = len(sys.argv) > 1
+
+    if not sim:
+        send_msg("ft8-txon")
+        time.sleep(1)
+        send_msg(zero)
+        time.sleep(3)
     
-    time.sleep(3)
-    
-    #freq_to_dac(g_tones[5])
+        #freq_to_dac(g_tones[5])
 
 
-    #f=open("g4rdc-cq-tones")
-    #cq_syms = [ int(x.strip()) for x in f.readlines() ]
-    #f.close()
-    recfile = '/home/john/ft8_t2.wav'
-    recfile_final = '/home/john/ft8_t2_sox.wav'
+        #f=open("g4rdc-cq-tones")
+        #cq_syms = [ int(x.strip()) for x in f.readlines() ]
+        #f.close()
+        recfile = '/home/john/ft8_t2.wav'
+        recfile_final = '/home/john/ft8_t2_sox.wav'
 
-    p = subprocess.Popen(['jack_capture', '-as', '--port', 'sdr_rx:ol', recfile ])
-    time.sleep(0.1)
+        p = subprocess.Popen(['jack_capture', '-as', '--port', 'sdr_rx:ol', recfile ])
+        time.sleep(0.1)
     
     #for i in test_syms3:
     for i in cq_syms:
         d = sym_to_dac(i)
-        send_dac(d)
-        time.sleep(0.159)
-        #time.sleep(0.158)
+        if sim:
+            (c, v) = d
+            if c==4:
+                print v-0.5
+            else:
+                print v
+        else:
+            send_dac(d)
+            #time.sleep(0.159)
+            time.sleep(0.159)
 
 
-    send_msg(zero)
+    if not sim:
+        send_msg(zero)
     
-    time.sleep(1)
+        time.sleep(1)
     
-    p.terminate()
+        p.terminate()
 
-    status = p.communicate()
-    print status
+        status = p.communicate()
+        print status
 
-    retcode = p.wait()
-    print retcode
+        retcode = p.wait()
+        print retcode
 
-    send_msg("ft8-txoff")
-    os.system(string.join(['sox', recfile, '-t wavpcm --rate 12k -c 1 -b 16', recfile_final, '&&', 'rm -f', recfile]))
+        send_msg("ft8-txoff")
+        os.system(string.join(['sox', recfile, '-t wavpcm --rate 12k -c 1 -b 16', recfile_final, '&&', 'rm -f', recfile]))
 
-    #send_msg("D2C2E")
-    #send_msg("Q90")
-    #send_dac(1529)
+        os.system(string.join(['python', '/home/john/basicft8/basicft8.py', '/home/john/ft8_t2_sox.wav', '>', '/home/john/basicft8/x']))
+        #send_msg("D2C2E")
+        #send_msg("Q90")
+        #send_dac(1529)
